requirements.txt
google-adk
PyYAML
python-dotenv
google-cloud-bigquery
